+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * First Version
 
 
-**1.0.3 Milfork Update**
+**1.0.3 Millfork Update**
 * Improved Syntax Highlighting
+
+
+**1.1.0 Bug Fix/Toolkit Feature**
+* Fixes issue the stop [Options Include Directory] form working
+* Added Toolkit Download Command